refactor: Steps 클래스 제거
