+++ src/main/java/ladder/domain/step/Step.java
@@ -2,6 +2,7 @@ package ladder.domain.step;
 
 import ladder.domain.step.strategy.CrossStrategy;
 import ladder.domain.step.strategy.RandomCrossStrategy;
+import ladder.domain.step.strategy.UnCrossableStrategy;
 
 public class Step {
     private final CrossStrategy crossStrategy;
@@ -23,5 +24,17 @@ public class Step {
 
     public boolean isCrossable() {
         return crossStrategy.isCrossable();
+    }
+
+    public static Step first() {
+        return new Step();
+    }
+
+    public static Step from(Step previousStep) {
+        if (previousStep.isCrossable()) {
+            return new Step(new UnCrossableStrategy());
+        }
+
+        return new Step();
     }
 }
